// src/generated/prisma/enums.ts
export const contract_type_vehicles_enum = {
  Leasing: 'Leasing',
  Alquiler: 'Alquiler',
  Propio: 'Propio',
  Prendado: 'Prendado',
} as const;

export type contract_type_vehicles_enum =
  (typeof contract_type_vehicles_enum)[keyof typeof contract_type_vehicles_enum];
// ...
export const daily_report_status = {
  pendiente: 'pendiente',
  sin_recursos_asignados: 'sin_recursos_asignados',
  ejecutado: 'ejecutado',
  reprogramado: 'reprogramado',
  cancelado: 'cancelado',
  en_certificacion: 'en_certificacion',
} as const;

export type daily_report_status = (typeof daily_report_status)[keyof typeof daily_report_status];
// ...
export const repair_state = {
  Pendiente: 'Pendiente',
  Esperando_repuestos: 'Esperando_repuestos',
  En_reparaci_n: 'En_reparaci_n',
  Finalizado: 'Finalizado',
  Rechazado: 'Rechazado',
  Cancelado: 'Cancelado',
  Programado: 'Programado',
} as const;

export type repair_state = (typeof repair_state)[keyof typeof repair_state];
// ...
export const work_order_status = {
  pending: 'pending',
  in_progress: 'in_progress',
  paused: 'paused',
  completed: 'completed',
  completed_partial: 'completed_partial',
  cancelled: 'cancelled',
} as const;

export type work_order_status = (typeof work_order_status)[keyof typeof work_order_status];